+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `--depclean/-c` flag: to perform 'emerge --depclean' before each test install
+
 ## [0.2.7] - 2025-11-24
 
 ### Added